+++ pysaql/saql/scalar.py
@@ -309,3 +309,23 @@ class Scalar(Expression, BooleanOperation, ABC):
 
         """
         return BinaryOperation(operator.contains, self, iterable)
+
+
+class field(Scalar):
+    """Represents a field (column) in the data stream"""
+
+    name: str
+
+    def __init__(self, name: str) -> None:
+        """Initializer
+
+        Args:
+            name: Name of the field
+
+        """
+        super().__init__()
+        self.name = name
+
+    def to_string(self) -> str:
+        """Cast the field to a string"""
+        return escape_identifier(self.name)
